+++ src/components/WordGroupSelector/WordGroupIcon/WordGroupIcon.css
