+++ src/components/WordGroupSelector/WordGroupIcon/WordGroupIcon.js
